requirements.txt
Flask==3.1.2
pymongo==4.16.0
paho-mqtt==1.6.1
Jinja2==3.1.6
Werkzeug==3.1.5
dnspython==2.8.0